+++ StripeIdentity/StripeIdentityTests/Unit/NativeComponents/Coordinators/IdentityVerificationSheetTest.swift
@@ -27,9 +27,7 @@ final class IdentityVerificationSheetTest: XCTestCase {
     func testInvalidSecret() {
         var result: IdentityVerificationSheet.VerificationFlowResult?
         let sheet = sheetWithWebUI(clientSecret: "bad secret")
-        // TODO(mludowise|RUN_MOBILESDK-120): Using `presentInternal` instead of
-        // `present` so we can run tests on our CI until it's updated to iOS 14.
-        sheet.presentInternal(from: mockViewController) { (r) in
+        sheet.present(from: mockViewController) { (r) in
             result = r
         }
         guard case let .flowFailed(error) = result else {
@@ -54,9 +52,7 @@ final class IdentityVerificationSheetTest: XCTestCase {
 
     func testAnalytics() {
         let sheet = sheetWithWebUI()
-        // TODO(mludowise|RUN_MOBILESDK-120): Using `presentInternal` instead of
-        // `present` so we can run tests on our CI until it's updated to iOS 14.
-        sheet.presentInternal(from: mockViewController) { _ in }
+        sheet.present(from: mockViewController) { _ in }
 
         // Verify presented analytic is logged
         XCTAssertEqual(mockAnalyticsClient.loggedAnalytics.count, 1)
@@ -92,7 +88,7 @@ final class IdentityVerificationSheetTest: XCTestCase {
         )
         let sheet = sheetWithWebUI()
 
-        sheet.presentInternal(from: mockPresentingViewController) { _ in
+        sheet.present(from: mockPresentingViewController) { _ in
             exp.fulfill()
         }
         sheet.verificationFlowWebViewController(mockWebViewController, didFinish: .flowCanceled)
@@ -104,7 +100,7 @@ final class IdentityVerificationSheetTest: XCTestCase {
         let mockPresentingViewController = UIViewController(nibName: nil, bundle: nil)
         let sheet = sheetWithNativeUI()
 
-        sheet.presentInternal(from: mockPresentingViewController) { _ in
+        sheet.present(from: mockPresentingViewController) { _ in
             exp.fulfill()
         }
         sheet.verificationSheetController(mockVerificationSheetController, didFinish: .flowCanceled)