+++ Tests/installation_tests/carthage/CarthageTest/ViewController.swift
@@ -19,7 +19,11 @@ class ViewController: UIViewController {
         super.viewDidLoad()
         StripeAPI.defaultPublishableKey = "test"
         StripeAPI.paymentRequest(withMerchantIdentifier: "test", country: "US", currency: "USD")
-        let _ = IdentityVerificationSheet(verificationSessionClientSecret: "test")
+
+        if #available(iOS 14.3, *) {
+            let _ = IdentityVerificationSheet(verificationSessionClientSecret: "test")
+        }
+
         if #available(iOS 12.0, *) {
             let _ = ConnectionsSheet(linkAccountSessionClientSecret: "")
         }