+++ Tests/installation_tests/cocoapods/with_frameworks_swift/CocoapodsTest/ViewController.swift
@@ -18,7 +18,11 @@ class ViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         StripeAPI.defaultPublishableKey = "test"
-        let _ = IdentityVerificationSheet(verificationSessionClientSecret: "test")
+
+        if #available(iOS 14.3, *) {
+            let _ = IdentityVerificationSheet(verificationSessionClientSecret: "test")
+        }
+
         if #available(iOS 12.0, *) {
             let _ = ConnectionsSheet(linkAccountSessionClientSecret: "")
         }