+++ Tests/installation_tests/swift_package_manager/with_swift/SPMTest/ViewController.swift
@@ -18,10 +18,15 @@ class ViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         StripeAPI.defaultPublishableKey = "foo"
-        let _ = IdentityVerificationSheet(verificationSessionClientSecret: "test")
+
+        if #available(iOS 14.3, *) {
+            let _ = IdentityVerificationSheet(verificationSessionClientSecret: "test")
+        }
+
         if #available(iOS 12.0, *) {
             let _ = ConnectionsSheet(linkAccountSessionClientSecret: "")
         }
+
         // Initialize a card field to make sure we can load image resources
         let cardField = STPPaymentCardTextField()
         cardField.frame = CGRect(x: 0, y: 0, width: 300, height: 100)